Rename request class and add failure tracking
